+++ ruma-identifiers-validation/src/signing_key_id.rs
@@ -1,12 +1,12 @@
 use std::{num::NonZeroU8, str::FromStr};
 
-use crate::{crypto_algorithms::ServerKeyAlgorithm, Error};
+use crate::{crypto_algorithms::SigningKeyAlgorithm, Error};
 
 pub fn validate(s: &str) -> Result<NonZeroU8, Error> {
-    let colon_idx = NonZeroU8::new(s.find(':').ok_or(Error::MissingServerKeyDelimiter)? as u8)
+    let colon_idx = NonZeroU8::new(s.find(':').ok_or(Error::MissingSigningKeyDelimiter)? as u8)
         .ok_or(Error::UnknownKeyAlgorithm)?;
 
-    validate_server_key_algorithm(&s[..colon_idx.get() as usize])?;
+    validate_signing_key_algorithm(&s[..colon_idx.get() as usize])?;
     validate_version(&s[colon_idx.get() as usize + 1..])?;
 
     Ok(colon_idx)
@@ -22,8 +22,8 @@ fn validate_version(version: &str) -> Result<(), Error> {
     Ok(())
 }
 
-fn validate_server_key_algorithm(algorithm: &str) -> Result<(), Error> {
-    match ServerKeyAlgorithm::from_str(algorithm) {
+fn validate_signing_key_algorithm(algorithm: &str) -> Result<(), Error> {
+    match SigningKeyAlgorithm::from_str(algorithm) {
         Ok(_) => Ok(()),
         Err(_) => Err(Error::UnknownKeyAlgorithm),
     }