+++ ruma-identifiers/src/lib.rs
@@ -28,13 +28,13 @@ pub use crate::{
     room_id::RoomId,
     room_id_or_room_alias_id::RoomIdOrAliasId,
     room_version_id::RoomVersionId,
-    server_key_id::ServerKeyId,
     server_name::{ServerName, ServerNameBox},
+    signing_key_id::SigningKeyId,
     user_id::UserId,
 };
 #[doc(inline)]
 pub use ruma_identifiers_validation::{
-    crypto_algorithms::{DeviceKeyAlgorithm, EventEncryptionAlgorithm, ServerKeyAlgorithm},
+    crypto_algorithms::{DeviceKeyAlgorithm, EventEncryptionAlgorithm, SigningKeyAlgorithm},
     error::Error,
 };
 
@@ -50,8 +50,8 @@ mod room_alias_id;
 mod room_id;
 mod room_id_or_room_alias_id;
 mod room_version_id;
-mod server_key_id;
 mod server_name;
+mod signing_key_id;
 
 /// Check whether a given string is a valid server name according to [the specification][].
 ///
@@ -137,7 +137,7 @@ macro_rules! room_version_id {
     };
 }
 
-/// Compile-time checked `ServerKeyId` construction.
+/// Compile-time checked `SigningKeyId` construction.
 #[macro_export]
 macro_rules! server_key_id {
     ($s:literal) => {